air_ML/requirements.txt
# Core Vision
opencv-python>=4.8.0
numpy>=1.24.0
deepface>=0.0.79
dlib>=19.24.0

# Core Audio
sounddevice>=0.4.6
pyaudio>=0.2.13
wave>=0.0.2
pvporcupine>=2.1.0
speechbrain>=0.5.15
librosa>=0.10.0
gtts>=2.3.2
playsound>=1.3.0
SpeechRecognition>=3.10.0
pygame>=2.6.1
tf-keras>=2.15.0

# Machine Learning
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
transformers>=4.30.0

# API & Web
openai==0.28.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
scipy>=1.10.0